refactor: move postCancelModerationAndRedirectControllerFactory into edit-moderation.ts file
