+++ src/routers/user.ts
@@ -5,14 +5,16 @@ import {
   getJoinOrganizationController,
   getOrganizationSuggestionsController,
   getUnableToAutoJoinOrganizationController,
-  postCancelModerationAndRedirectControllerFactory,
   postJoinOrganizationMiddleware,
   postQuitUserOrganizationController,
 } from "../controllers/organization";
 import { postSignInWithAuthenticatorAppController } from "../controllers/totp";
 import { get2faSignInController } from "../controllers/user/2fa-sign-in";
 import { postDeleteUserController } from "../controllers/user/delete";
-import { getEditModerationController } from "../controllers/user/edit-moderation";
+import {
+  getEditModerationController,
+  postCancelModerationAndRedirectControllerFactory,
+} from "../controllers/user/edit-moderation";
 import { issueSessionOrRedirectController } from "../controllers/user/issue-session-or-redirect";
 import {
   getMagicLinkSentController,
